My first programs
The first two programs I've created in Python 3.x.

Dicey is a dice-rolling simulator and random number generator. You have the choice of rolling a six-sided dice, a D20 or using the random number generator to generate a number between a minimum and maximum that you specify yourself.
Guessy is a guess-the-number game.